refactor: split socket init to another utils script
